Improvements on Defiance Unit loading/rendering
vorador1a looks good now, but shold1a still has broken vertices.
Pillars, cemetary and elder god areas all seem ok.

+++ CDC/Models/SR1Model.cs
@@ -309,7 +309,7 @@ namespace CDC.Objects.Models
 
             xReader.BaseStream.Position = _vertexStart;
 
-            for (UInt16 v = 0; v < _vertexCount; v++)
+            for (int v = 0; v < _vertexCount; v++)
             {
                 ReadVertex(xReader, v);
             }
